[glue] Simplify stateful verification scheduling
