+++ glue/src/stateful/actor/processor/mod.rs
@@ -55,7 +55,10 @@ use std::{
     hash::Hash,
     sync::Arc,
 };
-use tracing::{debug, info_span, warn};
+use tracing::{debug, warn};
+
+mod verifier;
+pub(super) use verifier::Verifier;
 
 pub(super) type PendingDigest<A, E> = <<A as Application<E>>::Block as Digestible>::Digest;
 type PendingBatches<A, E> = <<A as Application<E>>::Databases as DatabaseSet<E>>::Merkleized;
@@ -586,31 +589,6 @@ where
     pruning: Option<Pruning<PendingSyncTargets<A, E>>>,
 }
 
-/// Executes one independently-polled verification request.
-pub(super) struct Verifier<E, A>
-where
-    E: Rng + Spawner + Metrics + Clock,
-    A: Application<E>,
-{
-    app: A,
-    execution: Execution<E, A>,
-    replays: ReplayFlights<PendingDigest<A, E>>,
-}
-
-impl<E, A> Clone for Verifier<E, A>
-where
-    E: Rng + Spawner + Metrics + Clock,
-    A: Application<E>,
-{
-    fn clone(&self) -> Self {
-        Self {
-            app: self.app.clone(),
-            execution: self.execution.clone(),
-            replays: self.replays.clone(),
-        }
-    }
-}
-
 impl<E, A> Processor<E, A>
 where
     E: Rng + Spawner + Metrics + Clock,
@@ -960,227 +938,6 @@ where
     ) -> bool {
         self.execution
             .cache_pending(digest, parent, round, merkleized)
-    }
-}
-
-impl<E, A> Verifier<E, A>
-where
-    E: Rng + Spawner + Metrics + Clock,
-    A: Application<E>,
-{
-    /// Verifies one request while allowing unrelated requests to be polled.
-    pub(super) async fn verify<S, V>(
-        &mut self,
-        context: &E,
-        marshal: MarshalMailbox<S, V>,
-        (runtime_context, consensus_context): (E, A::Context),
-        ancestry: impl Ancestry<A::Block>,
-        progress: &VerificationProgress<PendingDigest<A, E>>,
-        verification: &mut Verification,
-    ) -> Option<bool>
-    where
-        S: Scheme,
-        V: MarshalVariant<ApplicationBlock = A::Block>,
-        MarshalMailbox<S, V>: BlockProvider<Block = A::Block>,
-    {
-        let timer = self.execution.metrics.verify_duration.timer(context);
-        let mut ancestry = ancestry;
-
-        let block = match fetch_ancestor(verification, &mut ancestry).await {
-            Some(Some(block)) => block,
-            Some(None) => {
-                debug!("verification request waiting on incomplete block ancestry");
-                verification.cancelled().await;
-                return None;
-            }
-            None => {
-                debug!("verification request cancelled before initial block arrived");
-                return None;
-            }
-        };
-        let block_digest = block.digest();
-
-        if self.execution.pending_contains(&block_digest) {
-            timer.observe(context);
-            return Some(true);
-        }
-
-        // The voter may ask us to verify blocks that are at or below the
-        // already-processed height. This happens because marshal/state sync and
-        // simplex advance on different message streams.
-        //
-        // Re-execution is impossible because databases already contain state at
-        // or beyond that height, but we still need to prove the block matches
-        // the canonical finalized chain before short-circuiting.
-        //
-        // `last_processed.height` is only advanced from finalized state
-        // (genesis, startup reconciliation, or finalize/ack path).
-        let last_processed = self.execution.last_processed();
-        match is_already_processed(
-            last_processed,
-            marshal.clone(),
-            block.as_ref(),
-            verification,
-        )
-        .await
-        {
-            Ok(true) => {
-                timer.observe(context);
-                return Some(true);
-            }
-            Ok(false) => {
-                if block.height() <= last_processed.height {
-                    return Some(false);
-                }
-            }
-            Err(PrepareBatchesError::Cancelled) => {
-                debug!(
-                    ?block_digest,
-                    "verification request cancelled during processed-block check"
-                );
-                return None;
-            }
-            Err(PrepareBatchesError::Incomplete) => {
-                debug!(
-                    ?block_digest,
-                    "verification request waiting on incomplete processed-block ancestry"
-                );
-                verification.cancelled().await;
-                return None;
-            }
-            Err(PrepareBatchesError::Invalid) => {
-                unreachable!("processed-block check cannot return Invalid")
-            }
-        }
-
-        let round = consensus_context.round();
-        let parent = match fetch_ancestor(verification, &mut ancestry).await {
-            Some(Some(parent)) => parent,
-            Some(None) => {
-                debug!(
-                    ?block_digest,
-                    "verification request waiting on incomplete parent ancestry"
-                );
-                verification.cancelled().await;
-                return None;
-            }
-            None => {
-                debug!(
-                    ?block_digest,
-                    "verification request cancelled before parent ancestry arrived"
-                );
-                return None;
-            }
-        };
-        let parent_digest = parent.digest();
-        let batches = match self
-            .execution
-            .prepare_batches(
-                &mut self.app,
-                context,
-                marshal.clone(),
-                parent.clone(),
-                verification,
-                Some(ReplayTracking {
-                    flights: &self.replays,
-                    progress: Some(progress),
-                }),
-            )
-            .await
-        {
-            Ok(batches) => batches,
-            Err(PrepareBatchesError::Invalid) => {
-                let (last_processed, pending_keys) = self.execution.summary();
-                warn!(
-                    ?parent_digest,
-                    ?block_digest,
-                    pending_keys,
-                    last_processed = ?last_processed.digest,
-                    "verification rejected: prepare_batches returned Invalid"
-                );
-                return Some(false);
-            }
-            Err(PrepareBatchesError::Incomplete) => {
-                debug!(
-                    ?parent_digest,
-                    ?block_digest,
-                    "verification request waiting on incomplete ancestry during prepare_batches"
-                );
-                verification.cancelled().await;
-                return None;
-            }
-            Err(PrepareBatchesError::Cancelled) => {
-                debug!(
-                    ?parent_digest,
-                    "verification request cancelled during prepare_batches"
-                );
-                return None;
-            }
-        };
-
-        progress.verifying(block_digest, parent_digest, round);
-        let ancestry = marshal_ancestry::with_prefix([block.clone(), parent], ancestry);
-        let verified = match await_or_cancel(
-            verification,
-            self.app.verify(
-                (
-                    runtime_context.child("verify_attempt"),
-                    consensus_context.clone(),
-                ),
-                ancestry,
-                batches,
-            ),
-        )
-        .await
-        {
-            Some(result) => result,
-            None => {
-                debug!(
-                    ?parent_digest,
-                    "verification request cancelled during verify"
-                );
-                return None;
-            }
-        };
-
-        let Some(merkleized) = verified else {
-            warn!(
-                ?parent_digest,
-                ?block_digest,
-                "verification rejected: app.verify returned None"
-            );
-            return Some(false);
-        };
-        let tail = info_span!(
-            "stateful.processor.match_commitments",
-            block = %block_digest,
-            parent = %parent_digest,
-        )
-        .entered();
-        if !A::Databases::matches_sync_targets(&merkleized, &A::sync_targets(&block)) {
-            warn!(
-                ?parent_digest,
-                ?block_digest,
-                "verification rejected: verified state must match block commitments"
-            );
-            return Some(false);
-        }
-        if !self
-            .execution
-            .cache_pending(block_digest, parent_digest, round, merkleized)
-        {
-            warn!(
-                ?parent_digest,
-                ?block_digest,
-                "verification result became incompatible before publication"
-            );
-            return Some(false);
-        }
-        self.execution.update_pending_metric();
-        drop(block);
-        drop(tail);
-        timer.observe(context);
-        Some(true)
     }
 }
